Fixed Privacy API and Added Deactivate and Set account for deletion.

+++ backend/src/controllers/PostController.ts
@@ -7,7 +7,7 @@ import { storage } from "../config/appwrite";
 import { INTERNAL_SERVER_ERROR } from "../config/commonErrors";
 
 // Models
-import Post, {IPost} from "../models/post.model";
+import Post, { IPost } from "../models/post.model";
 import Like from "../models/like.model";
 import Comment from "../models/comment.model";
 import User from "../models/user.model";
@@ -16,58 +16,95 @@ import User from "../models/user.model";
 import { InputFile } from "node-appwrite/file";
 import mongoose, { mongo } from "mongoose";
 
-
-export const PublicPosts = async (req: Request, res: Response):Promise<void> => {
+export const PublicPosts = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
   try {
-    const posts:IPost[] = await Post.find({users: req.signedInUser?.id})  
+    const posts: IPost[] = await Post.find({
+      users: req.signedInUser?.id,
+      is_paid: false,
+    });
 
-    const publicPosts:IPost[] = posts.map((post) => {
-      if(!post.is_paid){
-        return post.toObject()
-      }
-    })  
+    const publicPosts: IPost[] = await Promise.all(
+      posts.map(async (post): Promise<IPost> => {
+        const likeCount = await Like.countDocuments({
+          post_user: post.users,
+        });
+        const commentCount = await Comment.countDocuments({
+          post_user: post.users,
+        });
 
-    if(!posts){
+        return {
+          ...post.toObject(),
+          likes_count: likeCount,
+          comment_count: commentCount,
+        };
+      })
+    );
+
+    if (!posts) {
       res.status(404).json({
         ok: true,
-        msg: "No Posts."
-      })
+        msg: "No Posts.",
+      });
       return;
     }
 
     res.status(200).json({
       ok: true,
-      msg: 'All public posts fetched.',
-      posts: publicPosts[0] != null?publicPosts:[]
-    })
+      msg: "All public posts fetched.",
+      posts: publicPosts[0] != null ? publicPosts : [],
+    });
   } catch (error) {
-    INTERNAL_SERVER_ERROR(res, error, "AllPost Controller.")
+    INTERNAL_SERVER_ERROR(res, error, "AllPost Controller.");
   }
-}
+};
 
-
-export const PaidPosts = async (req: Request, res: Response):Promise<void> => {
+export const PaidPosts = async (req: Request, res: Response): Promise<void> => {
   try {
-    const paidPosts:IPost[] = await Post.find({user: req.signedInUser?.id, is_paid: true})
+    const rawPaidPosts: IPost[] = await Post.find({
+      user: req.signedInUser?.id,
+      is_paid: true,
+    });
 
-    if(!paidPosts) {
+    const paidPosts: IPost[] = await Promise.all(
+      rawPaidPosts.map(async (post): Promise<IPost> => {
+        if (!post.is_paid) {
+          const likeCount = await Like.countDocuments({
+            post_user: post.users,
+          });
+          const commentCount = await Comment.countDocuments({
+            post_user: post.users,
+          });
+
+          return {
+            ...post.toObject(),
+            likes_count: likeCount,
+            comment_count: commentCount,
+          };
+        }
+        return post.toObject();
+      })
+    );
+
+    if (!paidPosts) {
       res.status(404).json({
         ok: true,
-        msg: "No Paid Posts."
-      })
+        msg: "No Paid Posts.",
+      });
       return;
     }
 
     res.json({
       ok: true,
       msg: "All paid posts.",
-      paid_posts: paidPosts
-    })
+      paid_posts: paidPosts,
+    });
   } catch (error) {
-    INTERNAL_SERVER_ERROR(res, error, "PaidPost")
+    INTERNAL_SERVER_ERROR(res, error, "PaidPost");
   }
-}
-
+};
 
 export const Posts = async (req: Request, res: Response): Promise<void> => {
   try {
@@ -116,11 +153,14 @@ export const Posts = async (req: Request, res: Response): Promise<void> => {
       return;
     }
 
-    await User.updateOne({_id: req.signedInUser?.id}, {
-      $inc:{
-        posts: 1
+    await User.updateOne(
+      { _id: req.signedInUser?.id },
+      {
+        $inc: {
+          posts: 1,
+        },
       }
-    })
+    );
 
     res.status(201).json({
       ok: true,
@@ -147,12 +187,22 @@ export const LikeUnlikePost = async (
       return;
     }
 
+    const post = await Post.findById(post_id);
+
+    if (!post) {
+      res.status(404).json({
+        ok: false,
+        msg: "Post not found.",
+      });
+      return;
+    }
+
     const isAlreadyLiked = await Like.findOne({
       post_id,
       liked_by: req.signedInUser?.id,
     });
     if (isAlreadyLiked) {
-      await Like.findOneAndDelete({post_id, liked_by: req.signedInUser?.id});
+      await Like.findOneAndDelete({ post_id, liked_by: req.signedInUser?.id });
       await Post.findByIdAndUpdate(post_id, {
         $inc: {
           like_count: -1,
@@ -166,6 +216,7 @@ export const LikeUnlikePost = async (
     } else {
       const isLiked = await Like.create({
         post_id,
+        post_user: post.users,
         liked_by: req.signedInUser?.id,
       });
 
@@ -193,176 +244,210 @@ export const LikeUnlikePost = async (
   }
 };
 
-
-export const PostComment = async (req: Request, res: Response):Promise<void> => {
+export const PostComment = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
   try {
-    const {post_id, comment} = req.query;
+    const { post_id, comment } = req.query;
 
-    if(!post_id || typeof post_id !== "string" || !comment || typeof comment !== "string" || !mongoose.isValidObjectId(post_id)){
+    if (
+      !post_id ||
+      typeof post_id !== "string" ||
+      !comment ||
+      typeof comment !== "string" ||
+      !mongoose.isValidObjectId(post_id)
+    ) {
       res.status(400).json({
         ok: false,
-        msg: 'PostId and Comment is required.'
-      })
+        msg: "PostId and Comment is required.",
+      });
       return;
     }
 
-    const isPost = await Post.findById(post_id)
-    if(!isPost){
+    const isPost = await Post.findById(post_id);
+    if (!isPost) {
       res.status(404).json({
         ok: false,
-        msg: "Post not found."
-      })
+        msg: "Post not found.",
+      });
       return;
     }
-    
+
     const isComment = await Comment.create({
       post_id,
+      post_user: isPost.users,
       comment_by: req.signedInUser?.id,
-      comment
-    })
+      comment,
+    });
 
-    if (!isComment){
+    if (!isComment) {
       res.status(500).json({
         ok: false,
-        msg:'Unable to post comment.'
-      })
+        msg: "Unable to post comment.",
+      });
       return;
     }
 
     await Post.findByIdAndUpdate(post_id, {
       $inc: {
-        comment_count: 1
-      }
-    })
+        comment_count: 1,
+      },
+    });
 
     res.status(201).json({
       ok: true,
-      msg: 'Comment posted.',
-      comment: isComment
-    })
+      msg: "Comment posted.",
+      comment: isComment,
+    });
   } catch (error) {
-    INTERNAL_SERVER_ERROR(res, error, "Comment Controller.")
+    INTERNAL_SERVER_ERROR(res, error, "Comment Controller.");
   }
-}
+};
 
-
-export const DeleteComment = async (req: Request, res: Response):Promise<void> => {
+export const DeleteComment = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
   try {
-    const {comment_id} = req.query;
+    const { comment_id } = req.query;
 
-    if(!comment_id || typeof comment_id !== 'string' || !mongoose.isValidObjectId(comment_id)){
+    if (
+      !comment_id ||
+      typeof comment_id !== "string" ||
+      !mongoose.isValidObjectId(comment_id)
+    ) {
       res.status(400).json({
         ok: false,
-        msg: 'CommentId is required.'
-      })
+        msg: "CommentId is required.",
+      });
       return;
     }
 
-    const isDeleted = await Comment.findByIdAndDelete(comment_id)
-    if(!isDeleted){
+    const isDeleted = await Comment.findByIdAndDelete(comment_id);
+    if (!isDeleted) {
       res.status(400).json({
         ok: false,
-        msg: 'Unable to delete comment.'
-      })
+        msg: "Unable to delete comment.",
+      });
       return;
     }
 
     await Post.findByIdAndUpdate(isDeleted?.post_id, {
       $inc: {
-        comment_count: -1
-      }
-    })
+        comment_count: -1,
+      },
+    });
 
     res.status(200).json({
       ok: true,
       msg: "Comment deleted.",
-    })
+    });
   } catch (error) {
-    INTERNAL_SERVER_ERROR(res, error, "Comment Controller.")
+    INTERNAL_SERVER_ERROR(res, error, "Comment Controller.");
   }
-}
+};
 
-export const DeletePost = async (req: Request, res: Response):Promise<void> => {
+export const DeletePost = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
   try {
-    const {post_id} = req.query;
+    const { post_id } = req.query;
 
-    const post = await Post.findById(post_id)
-    if(!post){
+    const post = await Post.findById(post_id);
+    if (!post) {
       res.status(404).json({
         ok: false,
-        msg: "Post not found."
-      })
+        msg: "Post not found.",
+      });
       return;
     }
 
-    if(!post?.users.includes(req.signedInUser?.id)){
+    if (!post?.users.includes(req.signedInUser?.id)) {
       res.status(401).json({
         ok: false,
-        msg: "You are not authorized to delete this post."
-      })
+        msg: "You are not authorized to delete this post.",
+      });
       return;
     }
 
-    const isLikeDeleted = await Like.deleteMany({post_id})
-    const isCommentDeleted = await Comment.deleteMany({post_id})
-    const isPostDeleted = await Post.deleteOne({_id: post_id})
+    const isLikeDeleted = await Like.deleteMany({ post_id });
+    const isCommentDeleted = await Comment.deleteMany({ post_id });
+    const isPostDeleted = await Post.deleteOne({ _id: post_id });
 
     const isUserUpdated = await User.findByIdAndUpdate(req.signedInUser?.id, {
       $inc: {
-        posts: -1
-      }
-    })
+        posts: -1,
+      },
+    });
 
-    const isFileDeleted = await storage.deleteFile(process.env.POST_BUCKET_ID as string, post.post_file_id[0] as string)
+    const isFileDeleted = await storage.deleteFile(
+      process.env.POST_BUCKET_ID as string,
+      post.post_file_id[0] as string
+    );
 
-    if(!isLikeDeleted || !isCommentDeleted || !isUserUpdated || !isPostDeleted || !isFileDeleted){
+    if (
+      !isLikeDeleted ||
+      !isCommentDeleted ||
+      !isUserUpdated ||
+      !isPostDeleted ||
+      !isFileDeleted
+    ) {
       res.status(400).json({
         ok: false,
-        msg: 'Unable to delete this post.'
-      })
+        msg: "Unable to delete this post.",
+      });
       return;
     }
 
     res.status(200).json({
       ok: true,
-      msg: "Post deleted successfully."
-    })
+      msg: "Post deleted successfully.",
+    });
   } catch (error) {
-    INTERNAL_SERVER_ERROR(res, error, "Delete Post")
+    INTERNAL_SERVER_ERROR(res, error, "Delete Post");
   }
-}
+};
 
-export const EditPost = async (req: Request, res: Response):Promise<void> => {
+export const EditPost = async (req: Request, res: Response): Promise<void> => {
   try {
-    const {post_id,caption} = req.query;
+    const { post_id, caption } = req.query;
 
-    if(typeof caption !== 'string' || !post_id || !mongoose.isValidObjectId(post_id)){
+    if (
+      typeof caption !== "string" ||
+      !post_id ||
+      !mongoose.isValidObjectId(post_id)
+    ) {
       res.status(400).json({
         ok: false,
-        msg:'Either post id is not given or caption.'
-      })
+        msg: "Either post id is not given or caption.",
+      });
       return;
     }
 
-    const isEdited = await Post.findOneAndUpdate({_id: post_id, users: req.signedInUser?.id}, {
-      $set: {
-        caption
+    const isEdited = await Post.findOneAndUpdate(
+      { _id: post_id, users: req.signedInUser?.id },
+      {
+        $set: {
+          caption,
+        },
       }
-    })
+    );
 
-    if(!isEdited){
+    if (!isEdited) {
       res.status(400).json({
         ok: false,
-        msg: 'Unable to edit post. Try again later.'
-      })
+        msg: "Unable to edit post. Try again later.",
+      });
       return;
     }
 
     res.json({
       ok: true,
-      msg: "Post edited."
-    })
+      msg: "Post edited.",
+    });
   } catch (error) {
-    INTERNAL_SERVER_ERROR(res,error, "EditPost Controller.")
+    INTERNAL_SERVER_ERROR(res, error, "EditPost Controller.");
   }
-}
+};
